requirements.txt
django
djangorestframework
python-social-auth
psycopg2
pillow
gunicorn
django-filter
messytables
